refactor: move readMenuInput to executor_run_input_mode.go
